Improve Couchbase auto-configuration
This commit improves the couchbase auto-configuration so that it is
easier to customize the way the connection to the couchbase server is
initiated.

See gh-11146
